feat: merge usability and red_flags into one file
